Split BlockRenderer class
